Move enums into new package.
