+++ docs/api_responses.md
@@ -599,8 +599,8 @@ These can be returned in response to multiple possible fields.
 These errors are specific to certain APIs. Only those APIs can return these errors.
 
 | Error Code | Field           | HTTP Code | APIs | Error Message                                                |
-|------------|-----------------|-----------|-----|--------------------------------------------------------------|
-| ltd00      | new_username    | 409       | [Register User](#register-user-user) [Change Username](#change-username-user) | New username already exists |
+|------------|-----------------|-----------|------|--------------------------------------------------------------|
+| ltd00      | username_hash   | 409       | [Register User](#register-user-user) [Change Username](#change-username-user) | New username already exists |
 | ltd01      | request_number  | 400       | [Delete User](#delete-user-user) | Request number must be 0 for this request type                   |
 | ltd02      | request         | 412       | [Complete Password Change](#complete-password-change-password) | Password change is not complete                     |
