+++ src/utils/__init__.py
@@ -2,4 +2,3 @@ from .logging_setup import setup_logging
 from .database_config import DatabaseConfig
 from .db_utils_user import DBUtilsUser
 from .service_utils import ServiceUtils
-from .utils_enums import FailureReason
