+++ src/utils/db_utils_auth.py
@@ -7,7 +7,7 @@ logger = getLogger("database")
 from sqlalchemy.orm import Session
 
 from database import DatabaseSetup, User, AuthEphemeral, LoginSession
-from .utils_enums import FailureReason
+from enums import FailureReason
 from .db_utils_password import DBUtilsPassword
